More leak fixes

+++ src/raytracer/scene/Scene.cpp
@@ -94,7 +94,7 @@ void Scene::loadMaterials(std::vector<raytracer::Material*> materials) {
 
 
 void Scene::loadTextures(std::vector<raytracer::Texture*> textures) {
-  for(int i = 0; i < textures.size();i++) {
+  for(size_t i = 0; i < textures.size();i++) {
     m_textures.push_back(textures.at(i));
   }
 }
